dataflow/fx_pipeline/requirements.txt
apache-beam[gcp]==2.62.0
requests>=2.24.0
google-cloud-pubsub>=2.13.0
google-cloud-storage>=2.19.0